Remove guildai dependency from main codebase
- Move print_guild_scalars (renamed to print_scalars), detect_remote, and load_secrets to common.py
- Remove origami/utils/guild.py (contained guild-dependent functions)
- Remove guildai from setup.py, pyproject.toml, requirements.txt, and uv.lock
- Update experiments/README.md to document guild removal and legacy experiment status
- All 190 tests pass successfully

Guild-specific functions (plot_scalar_history, get_run_*, etc.) were removed as they're no longer needed in the main codebase. Legacy experiments still use guild but are documented as such.
