+++ pyproject.toml
@@ -22,7 +22,6 @@ classifiers = [
 dependencies = [
     "click>=8.1.7",
     "click-option-group>=0.5.6",
-    "guildai>=0.9.0",
     "lightgbm>=4.5.0",
     "matplotlib>=3.9.2",
     "mdbrtools>=0.1.1",
@@ -36,6 +35,7 @@ dependencies = [
     "torch>=2.4.1",
     "tqdm>=4.66.4",
     "xgboost>=2.1.3",
+    "yanex",
 ]
 
 [project.optional-dependencies]
@@ -76,7 +76,7 @@ ignore = [
 ]
 
 [tool.pytest.ini_options]
-filterwarnings = [
-    # Suppress pkg_resources deprecation warning from guildai (deprecated dependency, will be removed)
-    "ignore:pkg_resources is deprecated:UserWarning:guild",
-]
+filterwarnings = []
+
+[tool.uv.sources]
+yanex = { path = "../yanex", editable = true }
