+++ requirements.txt
@@ -1,6 +1,5 @@
 click==8.1.7
 click-option-group==0.5.6
-guildai==0.9.0
 jupyter==1.1.1
 jupyter_contrib_nbextensions==0.7.0
 lightgbm==4.5.0
